+++ app/routes/components/Avatar/demos/Demo01.tsx
@@ -4,6 +4,7 @@ export default function Demo() {
   return (
     <Avatar size="2.5rem">
       <Avatar.Img src="https://github.com/peterlitszo.png" alt="Peterlits Zo" />
+      <Avatar.Fallback>PZ</Avatar.Fallback>
     </Avatar>
   );
 }