+++ src/components/Avatar/CHANGELOG.md
@@ -4,3 +4,5 @@
 - 2025-10-11: Using type `ColorName` for color prop.
 - 2025-10-21: Update avatar background color calculation for better contrast.
   And add a border.
+- 2025-11-18: Add `Avatar.Fallback` sub-component and support to show it when
+  `Avatar.Img` not loaded.